requirements.txt
pandas
numpy
matplotlib
pyarrow
scipy
scikit-learn
scikit-optimize
PyPortfolioOpt
python-binance